+++ mssp-20161228/core/Models/CreateServiceWorkOrderRequest.cs
@@ -134,6 +134,10 @@ namespace AlibabaCloud.SDK.Mssp20161228.Models
         [Validation(Required=false)]
         public string OwnerId { get; set; }
 
+        [NameInMap("Priority")]
+        [Validation(Required=false)]
+        public int? Priority { get; set; }
+
         /// <summary>
         /// <para>Start time.</para>
         /// <para>This parameter is required.</para>