+++ mssp-20161228/core/Properties/AssemblyInfo.cs
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("dfd75f10-7d1e-11f0-93f6-1d7c90dc88c6")]
+[assembly: Guid("42783f80-ced8-11f0-ad06-7b91fe0d8fba")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -30,4 +30,4 @@ using System.Runtime.InteropServices;
 //      Revision
 //
 [assembly: AssemblyVersion("1.3.2.0")]
-[assembly: AssemblyFileVersion("1.3.3.0")]
+[assembly: AssemblyFileVersion("1.3.4.0")]